Mandelbrot

Description

Module which get resolution of image and points and calculate correct color to point based on Mandelbrot algorithm

Documentation

drawMandelbrot

Arguments

:: Size
Resolution of generating image

-> Bound
Four coordinates of funtion bound

-> String
Name of generating file

-> IO ()
Result: Save file

Main function, invoke draw Mandelbrot set

runMandelbrotTests :: IO Counts

Run tests

runQuickCheckMandelbrot :: IO ()

Run quick tests