Plot

Description

Module which change resolution to array of points and draw the image based on colors from Mandelbrot

Documentation

type Coord

Arguments

= (Double, Double)
X, Y coordinate

Type of coordinate point

type Bound

Arguments

= (Coord, Coord)
Left top and right bottom bound

Type of bound the function

drawPlot

Arguments

:: (Coord -> Color)
Function converting a point to the appropriate color

-> Size
Resolution of generating image

-> Bound
Four coordinates of funtion bound

-> String
Name of generating file

-> IO ()
Result: Save file

Draw image

runPlotTests :: IO Counts

Run tests

runQuickCheckPlot :: IO ()

Run quick tests